requirements.txt
asgiref==3.5.2
Django==4.1.1
djangorestframework==3.14.0
psycopg==3.1.2
psycopg2==2.9.3
pytz==2022.2.1
sqlparse==0.4.3
typing-extensions==4.3.0
tzdata==2022.4
django-createsuperuserwithpassword==2.0.0
